fix(plex): extract filter helpers to shared module, fix TVDB fallback, and normalize filter inputs
